requirements.txt
Django==2.2.1
django-markdownx==2.0.28
gunicorn==19.9.0
Markdown==3.1.1
Pillow==6.0.0
psycopg2==2.8.2
pytz==2019.1
setproctitle==1.1.10
sqlparse==0.3.0
